pyproject.toml
[tool.stela]
default_environment = "test"
config_file_extension = "TOML"

[tool.pytest]
addopts = '--pydocstyle'

[tool.black]
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \venv
    | \.aws-sam
    | _build
    | buck-out
    | build
    | dist
    | alembic
    | node_modules
  )/
)
'''

[tool.poetry]
name = "noverde-cookiecutter"
version = "0.1.0"
description = "This project contains all boilerplate you need to create a new microservice."
authors = ["Noverde <devs@noverde.com.br>"]

[tool.poetry.dependencies]
python = "^3.7"
arrow = "*"
aws_lambda_context = "*"
loguru = "*"
marshmallow = "*"
marshmallow_enum = "*"
marshmallow-sqlalchemy = "*"
marshmallow-pynamo-db = "*"
python-json-logger = "*"
pynamodb = "4.3.0"
psycopg2 = "*"
sentry-sdk = "*"
SQLAlchemy = "*"
stela = "*"

[tool.poetry.dev-dependencies]
alembic = "*"
aws-sam-cli = {version="*", markers="platform_system != 'Windows'"}
black = {version = "*", allow-prereleases = true}
factory_boy = "*"
freezegun = "*"
ipdb = "*"
pre-commit = "*"
pytest-black = "*"
pytest-cov = "*"
pytest-flake8 = "*"
pytest-mock = "*"
pytest-mypy = "*"
pytest-postgresql = "*"
pytest-pydocstyle = "*"
pytest-watch = "*"
pytest-xdist ="*"
wheel = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"